Rough code for D-Bus
'dbus-test' is the server side or the sender side of the d-bus. It exposes a
getMap() method inside it, to the D-Bus. Essentially it enables the client
application to get the map of data from the server app.

The 'dbus' contains just one file, that makes a connection to the interface I
created earlier. It gets the QVariantMap that the other application has and prints it

The server app(dbus-test), thus obviously must be running while we are testing the client side(dbus)